requirements.txt
flask
flask_socketio
python-dotenv
azure-ai-projects
azure-identity
